apps/analytics/pyproject.toml
[project]
name = "snaptrash-analytics"
version = "0.1.0"
description = "Stage 6: Dryad MSW load + SQL aggregation + Prophet forecast + threshold detection"
requires-python = ">=3.11"
dependencies = [
    "databricks-sdk>=0.40.0",
    "databricks-sql-connector>=3.4",
    "pandas>=2.2",
    "numpy>=1.26",
    "prophet>=1.1.5",
    "mlflow>=2.16",
    "httpx>=0.27",
    "firecrawl-py>=1.5",
    "pyarrow>=17.0",
    "snaptrash-common",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]

[tool.uv.sources]
snaptrash-common = { path = "../../packages/common", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/snaptrash_analytics"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
